Move the rest of types.
