+++ libs/safekeeper_api/src/send_wal.rs
@@ -1,0 +1,2 @@
+//! This module contains structs used in safekeeper send_wal.rs which are also
+//! exposed in http API.
