+++ libs/safekeeper_api/src/state.rs
@@ -1,1 +1,3 @@
-
+//! This module contains structs which are exposed in the http API but at the
+//! same time part of the safekeeper control file. Any change of in them must be
+//! accompanied with control file upgrade.
